+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## [unreleased]
 
-- **IMPROVED**: Various performance improvements, incl. lazy-loading the library for search highlighting
-- **IMPROVED**: Creating a single minified production bundle
+- **IMPROVED**: Various performance improvements:
+  - Creating a single minified production bundle, should speed up initial loading time a lot when IO is slow
+  - lazy-loading the library for search highlighting, further init load speedup
+  - Various smaller improvements that increase search and render performance
+- **FIXED**: Bug in scoring that lastVisited: 0 was not treated as most recent (max score)
+- **FIXED**: Bug in scoring
+- **IMPROVED**: Improved unit-test coverage and switched for E2E testing to Playwright, leading to faster and less brittle tests.
 
 ## [v1.14.0]